cloud_functions/adjlist_to_bq/requirements.txt
functions-framework==3.*
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.4
networkx==3.1
pandas==1.4.3
pyarrow==15.0.2